Extracted router.
